Validate api-job against sds-operations
Fix pep8 errors
